modify folder structure
