+++ indexer/.gitignore
@@ -1,4 +1,1 @@
-__pycache__
-_docker
-.env
-venv
+.apibara
