+++ indexer/apibara-script.sh
