+++ indexer/graphql/.gitignore
@@ -1,0 +1,4 @@
+__pycache__
+_docker
+.env
+venv
