+++ indexer/graphql/pyproject.toml
