+++ indexer/graphql/src/indexer/utils.py
