+++ indexer/master_script.sh
