+++ indexer/scripts/run_adventurers.sh
