+++ indexer/scripts/run_battles.sh
