+++ indexer/scripts/run_beasts.sh
