+++ indexer/scripts/run_discoveries.sh
