+++ indexer/scripts/run_scores.sh
