+++ indexer/src/utils/encode.ts
