--- new-indexer/.gitignore
@@ -1,1 +1,0 @@
-.apibara
